+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Released on
 Contributors:
 - Matt Tankersley
 
+### 🚀 Added
+- add individual functions for getting grid region, spacing, and registration
+
 ### 🐛 Fixed
 - fix broken links for ice velocity datasets